Power Point presentation/raw SQL code
I uploaded my Power Point presentation as a PDF. I also included the raw SQL code for the project in a word doc and tried formatting into a text doc. The instruction says for the file naming convention ending needs to = 'code.sql'

I have done so and as a result, I received the above (3) files. They should be the same, but I went ahead and included them. I was unsure whether or not either of the files were incorrect/correct.

PLEASE LET ME KNOW IF ANYTHING IS OFF! :) 

Thanks so much and this was a great learning experience!